refactor: move logic to helper function
